organize types: Order / Report
