+++ types/Order/OrderCreateParameters.d.ts
@@ -1,0 +1,14 @@
+import { IAddress } from '../Address';
+import { Shipment } from '../Shipment';
+
+export declare interface IOrderCreateParameters {
+  reference?: string;
+  to_address: IAddress | string;
+  from_address: IAddress | string;
+  shipments: Shipment[];
+
+  /**
+   * optional array of ids that begin with "ca_"
+   */
+  carrier_accounts?: string[];
+}
