+++ types/Report/ReportObjectType.d.ts
@@ -1,0 +1,7 @@
+export declare type TReportObjectType =
+  | 'CashFlowReport'
+  | 'PaymentLogReport'
+  | 'RefundReport'
+  | 'ShipmentReport'
+  | 'ShipmentInvoiceReport'
+  | 'TrackerReport';
