+++ types/Report/index.d.ts
@@ -1,0 +1,2 @@
+export * from './Report';
+export * from './ReportObjectType';
