+++ types/index.d.ts
@@ -1,133 +1,15 @@
-import { Address, IAddress } from './Address';
-import { IBaseObject, IDatedObject, IObjectWithId } from './base';
+import { Address } from './Address';
+import { IBaseObject, IObjectWithId } from './base';
 import { Batch } from './Batch';
 import { CarrierAccount } from './CarrierAccount';
 import { Insurance } from './Insurance';
 import { Parcel } from './Parcel';
 import { Pickup } from './Pickup';
-import { IMessage, IRate, IShipment, ServiceLevel, Shipment } from './Shipment';
-import { Carrier, Tracker } from './Tracker';
+import { Shipment } from './Shipment';
+import { Tracker } from './Tracker';
 import { Event } from './Event';
 import { Webhook } from './Webhook';
-
-/**
- * The Order object represents a collection of packages and can be used for Multi-Piece Shipments.
- * Like a single Shipment each Order consists of a "to" and "from" Address to be used for each Shipment within the Order.
- * These Addresses will be copied to each Shipment so there is no need to specify them multiple times.
- * Each Shipment must then specify its Parcel, Options, and CustomsInfo.
- *
- * An Order created with valid Address Objects and Parcel data nested within the Order's Shipment object will automatically retrieve available shipping Rate options.
- *
- * @see https://www.easypost.com/docs/api/node#order-object
- */
-export declare interface IOrder extends IObjectWithId<'Order'>, IDatedObject {
-  /**
-   * An optional field that may be used in place of id in other API endpoints
-   */
-  reference?: string;
-
-  /**
-   * The destination address
-   */
-  to_address: IAddress;
-
-  /**
-   * The origin address
-   */
-  from_address: IAddress;
-
-  /**
-   * The shipper's address, defaults to from_address
-   */
-  return_address: IAddress;
-
-  /**
-   * The buyer's address, defaults to to_address
-   */
-  buyer_address: IAddress;
-
-  /**
-   * All associated Shipment objects. Maximum of 100.
-   */
-  shipments: IShipment[];
-
-  /**
-   * All associated Rate objects
-   */
-  rates: IRate[];
-
-  /**
-   * Any carrier errors encountered during rating
-   */
-  messages: IMessage[];
-
-  /**
-   * Set true to create as a return
-   */
-  is_return: boolean;
-}
-
-export declare type TReportObject =
-  | 'CashFlowReport'
-  | 'PaymentLogReport'
-  | 'RefundReport'
-  | 'ShipmentReport'
-  | 'ShipmentInvoiceReport'
-  | 'TrackerReport';
-
-/**
- * A Report contains a csv that is a log of all the objects created within a certain time frame.
- *
- * Reports can be generated using the Reports Endpoint.
- * You can create and view Reports created between any time frame defined between the start_date and end_date.
- *
- * The Report api can be categorized into several types.
- * These types determine which EasyPost Object to produce a Report for, and should be passed as the type in our libraries:
- *  - payment_log
- *  - refund
- *  - shipment
- *  - shipment_invoice
- *  - tracker
- *
- * @see https://www.easypost.com/docs/api/node#report-object
- */
-export declare interface IReport extends IObjectWithId<TReportObject>, IDatedObject {
-  /**
-   * "new", "available", "failed", or null
-   */
-  status: 'new' | 'available' | 'failed' | null;
-
-  /**
-   * A date string in YYYY-MM-DD form eg: "2016-02-02"
-   */
-  start_date: string;
-
-  /**
-   * A date string in YYYY-MM-DD form eg: "2016-02-03"
-   */
-  end_date: string;
-
-  /**
-   * Set true if you would like to include Refunds /Shipments /Trackers created by child users
-   */
-  include_children: boolean;
-
-  /**
-   * A url that contains a link to the Report.
-   * Expires 30 seconds after retrieving this object
-   */
-  url: string;
-
-  /**
-   * Url expiring time
-   */
-  url_expires_at: string;
-
-  /**
-   * Set true if you would like to send an email containing the Report
-   */
-  send_email: boolean;
-}
+import { Order } from './Order';
 
 /**
  * The CarrierType object provides an export declare interface for determining the valid fields of a CarrierAccount.
@@ -258,67 +140,6 @@ export declare interface IUser extends IObjectWithId<'User'> {
   children: IUser[];
 }
 
-export declare interface IOrderCreateParameters {
-  reference?: string;
-  to_address: IAddress | string;
-  from_address: IAddress | string;
-  shipments: Shipment[];
-
-  /**
-   * optional array of ids that begin with "ca_"
-   */
-  carrier_accounts?: string[];
-}
-
-export declare class Order implements IOrder {
-  public constructor(input: IOrderCreateParameters);
-
-  reference?: string;
-  to_address: IAddress;
-  from_address: IAddress;
-  return_address: IAddress;
-  buyer_address: IAddress;
-  shipments: IShipment[];
-  rates: IRate[];
-  messages: IMessage[];
-  is_return: boolean;
-  id: string;
-  mode: 'test' | 'production';
-  object: 'Order';
-  created_at: string;
-  updated_at: string;
-
-  /**
-   * An Order is almost exclusively a container for other objects, and thus an Order may reuse many of these objects.
-   * Alternatively, all the objects contained within an Order may be created at the same time.
-   *
-   * You can limit the CarrierAccounts to use for rating by passing the carrier_accounts parameter.
-   *
-   * @see https://www.easypost.com/docs/api/node#create-an-order
-   */
-  public save(): Promise<Order>;
-
-  /**
-   * An Order can be retrieved by either its id or reference.
-   * However it is recommended to use EasyPost's provided identifiers because uniqueness on reference is not enforced.
-   *
-   * @param orderId Unique, begins with "order_"
-   *
-   * @see https://www.easypost.com/docs/api/node#retrieve-an-order
-   */
-  static retrieve(orderId: string): Promise<Order>;
-
-  /**
-   * To purchase an Order you only need to specify the carrier and service to purchase.
-   * This operation populates the tracking_code and postage_label attributes of each Shipment.
-   *
-   * @see https://www.easypost.com/docs/api/node#buy-an-order
-   */
-  public buy(carrier: Carrier, service: ServiceLevel): Promise<Order>;
-
-  public getRates(): Promise<IRate[]>;
-}
-
 export declare class CarrierType implements ICarrierType {
   type: string;
   fields: ICarrierTypeFields;
@@ -346,6 +167,8 @@ export declare class Easypost {
   public Webhook: typeof Webhook;
   public CarrierType: typeof CarrierType;
   public Event: typeof Event;
+
+  // TODO Report class
 
   public constructor(apiKey: string);
 }
